Fix projects with remote mirrors asynchronously destruction (using pending_delete)
When we update the flag pending_delete to true and later try to 
destroy the remote mirror (through association callbacks) the remote mirror try to load the project, but it is out of the default scope so it just return nil.

As we need the project to remove the reference to the remote mirror but we’re going to delete the project and its repository is not needed to touch the repository to just remove that reference, so we just skip that step in that case

+++ app/models/remote_mirror.rb
@@ -158,7 +158,9 @@ class RemoteMirror < ActiveRecord::Base
   end
 
   def remove_remote
-    project.repository.remove_remote(ref_name)
+    if project # could be pending to delete so don't need to touch the git repository
+      project.repository.remove_remote(ref_name)
+    end
   end
 
   def mirror_url_changed?
